refactor + fix copyright
